data/requirement.txt
cassandra-driver==3.3.0
